Prueba falla en Lint

+++ eslint.config.js
@@ -6,6 +6,6 @@ export default {
   },
   rules: {
     semi: ['error', 'always'],
-    quotes: ['error', 'single'],
+    quotes: ['error', 'double'],
   },
 };
